requirements.txt
BeautifulReport==0.1.3
captcha==0.3
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
Django==2.2.4
django-haystack==2.6.1
django-pure-pagination==0.3.0
gevent==20.9.0
greenlet==0.4.17
idna==2.10
jieba==0.39
Markdown==2.6.11
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
pycparser==2.20
pyDes==2.0.1
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2017.3
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
zope.event==4.5.0
zope.interface==5.2.0
